Use conditional compilation blocks to identify Linux incompatibilities
Use `#if os(Linux) #else <code> #endif` to identify code that is incompatible with Linux.
